fix(import): make statement import transactional, extract StatementImportConfig, invert null guard
