+++ FinanceBackEnd/src/Finance.Application/Helpers/ExcelHelper/StatementImportExcelHelper.cs
@@ -1,5 +1,4 @@
 using ExcelDataReader;
-using Finance.Application.Commands.CreditCards;
 using Microsoft.AspNetCore.Http;
 
 namespace Finance.Helpers.ExcelHelper;
